re-organizing into bandwidth service. re-enable rollup loop
