+++ storagenode/peer.go
@@ -131,6 +131,8 @@ type Peer struct {
 		Service  *console.Service
 		Endpoint *consoleserver.Server
 	}
+
+	Bandwidth *bandwidth.Service
 }
 
 // New creates a new Storage Node.
@@ -321,6 +323,8 @@ func New(log *zap.Logger, full *identity.FullIdentity, db DB, config Config, ver
 
 	peer.Collector = collector.NewService(peer.Log.Named("collector"), peer.Storage2.Store, peer.DB.PieceInfo(), peer.DB.UsedSerials(), config.Collector)
 
+	peer.Bandwidth = bandwidth.NewService(peer.Log.Named("bandwidth"), peer.DB.Bandwidth())
+
 	return peer, nil
 }
 
@@ -354,9 +358,9 @@ func (peer *Peer) Run(ctx context.Context) (err error) {
 		return errs2.IgnoreCanceled(peer.Vouchers.Run(ctx))
 	})
 
-	//group.Go(func() error {
-	//	return errs2.IgnoreCanceled(peer.DB.Bandwidth().Run(ctx))
-	//})
+	group.Go(func() error {
+		return errs2.IgnoreCanceled(peer.Bandwidth.Run(ctx))
+	})
 
 	group.Go(func() error {
 		// TODO: move the message into Server instead
@@ -387,9 +391,9 @@ func (peer *Peer) Close() error {
 
 	// close services in reverse initialization order
 
-	//if peer.DB.Bandwidth() != nil {
-	//	errlist.Add(peer.DB.Bandwidth().Close())
-	//}
+	if peer.DB.Bandwidth() != nil {
+		errlist.Add(peer.Bandwidth.Close())
+	}
 	if peer.Vouchers != nil {
 		errlist.Add(peer.Vouchers.Close())
 	}
